# Use golang etcd client