--- a/glide.yaml
+++ b/glide.yaml
@@ -1,3 +1,9 @@
 package: github.com/dtan4/paus-frontend
 import:
 - package: github.com/gin-gonic/gin
+- package: github.com/coreos/etcd
+  subpackages:
+  - client
+- package: golang.org/x/net
+  subpackages:
+  - context