--- a/templates/urls.tmpl
+++ b/templates/urls.tmpl
@@ -14,7 +14,7 @@
       </div>
       {{ else }}
       <div class="list-group">
-        {{ range .appURLs }}<a class="list-group-item" href="{{ . }}" target="_blank">{{ . }}</a>{{ else }}Not Found URL{{ end }}
+        {{ range .urls }}<a class="list-group-item" href="{{ . }}" target="_blank">{{ . }}</a>{{ else }}Not Found URL{{ end }}
       {{ end }}
       </div>
     </div>