## WHY
Calling `etcd-ls.sh` consumes more time due to shell execution overhead. It also reduces maintainability.

## WHAT
[coreos/etcd/client](https://github.com/coreos/etcd/blob/master/client/keys.go) has API client compatible with `etcdctl ls`. Replace `etcd-ls.sh` to use golang native etcd client.

After this pull request, paus-frontend requires 2 or 3 environment variables.

- `BASE_DOMAIN`
  - Base domain for application, e.g. `pausapp.com`
- `ETCD_ENDPOINT`
  - Endpoint of etcd cluster, e.g. `http://hoge.com:2379`
- `URI_SCHEME` (optional, default: `http`)
  - URI scheme of application URL, `http` or `https`.

### Before 

```
[GIN] 2016/04/13 - 10:43:09 | 200 |   50.014057ms | ::1 |   GET     /urls
[GIN] 2016/04/13 - 10:43:12 | 200 |   44.102334ms | ::1 |   GET     /urls
```

### After

Approx. __5x faster__ :dash: 

```
[GIN] 2016/04/13 - 10:40:07 | 200 |   11.270455ms | ::1 |   GET     /urls
[GIN] 2016/04/13 - 10:40:10 | 200 |    6.370916ms | ::1 |   GET     /urls
```